last save

+++ projet 6/Back-end/controllers/stuff.js
@@ -25,33 +25,55 @@ exports.createThing = (req, res, next) => {
 };
 
 exports.modifyThing = async (req, res, next) => {
+    const token = req.headers.authorization.split(' ')[1];
+    const decodedToken = jwt.verify(token, process.env.TOKEN,);
+    const userId = decodedToken.userId;
+
+    /*  
+      if(typeof req.body === "object"){
+      }
+      if (!req.body.sauce || typeof req.body.sauce !== "string") {
+          return res.status(400).json({ error: "test22212" })
+      }
+*/
+    //const { name, manufacturer, description, mainPepper } = JSON.parse(req.body.sauce)
+    /* if (!req.body.name || !req.body.manufacturer || !req.body.description || !req.body.mainPepper) {
+         return res.status(400).json({ error: "test" })
+     }*/
+    //if (req.body.likes === -1 || req.body.likes === 0 || req.body.likes === 1) {
+    ;
     try {
         const checkSauce = await Thing.findOne({
             _id: req.params.id
         })
+        req.body.likes = checkSauce.likes
 
         if (checkSauce) {
-            const thingObject = req.file ? (
-                Thing.findOne({
-                    _id: req.params.id
-                }).then((sauce) => {
-                    const filename = sauce.imageUrl.split('/images/')[1]
-                    fs.unlinkSync(`images/${filename}`)
-                }),
-                {
-                    ...JSON.parse(req.body.sauce),
-                    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
-                }) : { ...req.body };
-            var opts = { runValidators: true };
-            console.log('------------------------------------');
-            console.log(thingObject);
-            console.log('------------------------------------');
-            //if (thingObject) {
-            //     return res.status(400).json({ error: "erreur la sauce n'a pas pu être modifié" })
-            //}
-            Thing.updateOne({ _id: req.params.id }, { ...thingObject, _id: req.params.id }, opts)
-                .then(() => res.status(200).json({ message: 'Objet modifié !' }))
-                .catch(error => res.status(400).json({ error: "erreur la sauce n'a pas pu être modifié" }))
+
+            if (userId === checkSauce.userId) {
+                const thingObject = req.file ? (
+                    Thing.findOne({
+                        _id: req.params.id
+                    }).then((sauce) => {
+                        const filename = sauce.imageUrl.split('/images/')[1]
+                        fs.unlinkSync(`images/${filename}`)
+                    }),
+                    {
+                        ...JSON.parse(req.body.sauce),
+                        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
+                    }) : { ...req.body };
+                var opts = { runValidators: true };
+
+
+                Thing.updateOne({ _id: req.params.id }, { ...thingObject, _id: req.params.id }, opts)
+                    .then(() => res.status(200).json({ message: 'Objet modifié !' }))
+                    .catch(error => res.status(400).json({ error: "erreur la sauce n'a pas pu être modifié" }))
+
+            } else {
+                res.status(400).json({ error: 'cette sauce ne vous appartient pas' })
+            }
+
+
         }
 
         else {
@@ -60,6 +82,11 @@ exports.modifyThing = async (req, res, next) => {
     } catch (error) {
         res.status(400).json({ error: 'une erreur est survenu' })
     }
+
+    // } else {
+
+    //     res.status(400).json({ error: '!!!!!' })
+    // }
 
 
 
@@ -87,10 +114,10 @@ exports.deleteThing = async (req, res, next) => {
                     })
                 //.catch(error => res.status(500).json({ error: 'cette sauce ne vous appartient pas' }));
             } else {
-                res.status(500).json({ error: 'cette sauce ne vous appartient pas' })
+                res.status(400).json({ error: 'cette sauce ne vous appartient pas' })
             }
         } else {
-            res.status(500).json({ error: 'cette sauce éxiste pas' })
+            res.status(400).json({ error: 'cette sauce éxiste pas' })
         }
     } catch (error) {
         res.status(400).json({ error: 'une erreur est survenu' })
